Sample project that illustrates the bug (with explicit/implicit switch)
The ability to switch between explicit animations (broken) and implicit animations (not broken) is convenient, but I'm afraid it makes the sample code more complicated that it should be. It's also not clear that the `.animation(nil)` modifier in the view tree doesn't break anything (in the explicit animation case).